chore : add exception handling

+++ coursemanager/src/main/java/com/nirdesh/coursemanager/repository/CourseMaterialRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CourseMaterialRepository extends JpaRepository<CourseMaterial,Integer> {
+public interface CourseMaterialRepository extends JpaRepository<CourseMaterial,Long> {
 
 }
